rename: 폴더명 snake_case로 변경
