+++ src/components/brewery_details/Carousel/Carousel.module.scss
