+++ src/components/brewery_details/Carousel/index.tsx
